+++ NEWS.md
@@ -1,5 +1,7 @@
 # sass (development version)
 
+- Closed #138: font_google(local = TRUE) now uses woff2 (instead of woff) for a font file type. (#139) 
+
 # sass 0.4.8
 
 - Closed #132: Fix R CMD check warning re error() format strings (for r-devel). (#133)