+++ NEWS.md
@@ -1,3 +1,5 @@
+# sass (development version)
+
 # sass 0.4.9
 
 - Closed #138: font_google(local = TRUE) now uses woff2 (instead of woff) for a font file type. (#139)